+++ src/redturtle/prenotazioni/utilities/dateutils.py
@@ -1,8 +1,9 @@
 # -*- coding: utf-8 -*-
+# from datetime import time
 from datetime import datetime
-from datetime import time
 from datetime import timedelta
 
+import pytz
 import six
 from plone.app.event.base import default_timezone
 
@@ -39,7 +40,10 @@ def hm2DT(day, hm, tzinfo=None):
         return None
     if len(hm) == 4 and ":" not in hm:
         hm = f"{hm[:2]}:{hm[2:]}"
-    return tzinfo.localize(datetime.combine(day, time.fromisoformat(hm)))
+    (h, m) = map(int, hm.split(":"))
+    return pytz.datetime.datetime(day.year, day.month, day.day, h, m, tzinfo=tzinfo)
+    # return tzinfo.localize(datetime.combine(day, time(h, m)))
+    # return tzinfo.localize(datetime.combine(day, time.fromisoformat(hm)))
 
 
 def hm2seconds(hm):