+++ src/redturtle/prenotazioni/utilities/dateutils.py
@@ -17,16 +17,13 @@ TIMEZONE_CACHE = True
 # NOTE: If the site timezone was changed, you need to reload the instance due to forever.memoize usage \cc @mamico
 def get_default_timezone(as_tzinfo):
     @forever.memoize
-    def cached_call():
-        return default_timezone(as_tzinfo=as_tzinfo)
-
-    def uncached_call():
+    def cached_call(as_tzinfo=True):
         return default_timezone(as_tzinfo=as_tzinfo)
 
     if TIMEZONE_CACHE:
-        return cached_call()
+        return cached_call(as_tzinfo=as_tzinfo)
     else:
-        return uncached_call()
+        return default_timezone(as_tzinfo=as_tzinfo)
 
 
 def hm2handm(hm):