+++ CHANGES.rst
@@ -14,6 +14,11 @@ Changelog
 - Add booking reminder emails.
   [folix-01]
 
+- Move code generation to adapter, to be more customizable.
+  [cekk]
+
+- Add customizable email from in PrenotazioniFolder contents.
+  [cekk]
 
 2.3.2 (2024-01-03)
 ------------------